Fix problem of dependency requirements in `setup_requires` of `pymatgen` (#3124)
The dependency `pymatgen` builds from source on any platform but MacOS
and include the requirement `numpy>=1.14` in its `setup_requires`. The
`setup_requires` rules are dealt with by `easy_install` which will
ignore the rules we put in our `install_requires`. On top of that, it is
not very strict with versioning and will install release candidates as
well. Since `numpy==1.17.1rc1` has been released, `easy_install` started
installing this, but this has dropped support for python 2, thus failing
all builds on that platform. The temporary fix is a hack by installing
`numpy==1.16.4` before we start our own install process.