Improved graph visualization (#2596)
The `python-graphviz` package is employed to simplify the code that can
build a visual representation of a sub set of the provenance graph. The
main graphing function is rewritten as a class, which:

 * maintains a cache of the added nodes and graphs (as sets)
 * splits up the function and allows for multiple additions to the graph
 * has separate `_style_` functions for different node classes

This gives larger control to define the graph that is to be visualized
as well as the style to be used. The current style is adapted to resemble
that of the documentation for a coherent look.