Up EXPORT_VERSION to 0.7
Added now illegal attributes (migration 0040) to process nodes in all
"migrate" fixture export archives.
Migration 0040 and the removal of 'Attribute' and 'Link' entities from
metadata.json and the import functions are the reason for the new
EXPORT_VERSION.

The dependency on aiida-export-migration-tests has been upped to 0.7.0,
together with the new release.

Due to migration 0040, it is now mandatory for ProcessNodes to be sealed
before they can be exported, hence this is checked in `export_tree`.
Because of this all tests including ProcessNodes has been updated.