Move responsibility of loading profile to `ProfileParamType` (#3138)
Currently, the loading of the profile is done in the body of the `verdi`
command group itself. By moving this to the `ProfileParamType`, it
becomes easy for external scripts to also provide a `--profile` flag to
run the script for a profile other than the default.

In addition, prevent loading the ORM backend through the configuration.
The `load_profile` method was calling `configure_logging` with the
option `with_orm=True` in order to make sure that the `DbLogHandler` was
properly configured for the new profile. However, `load_profile` should
not load the backend. This should be done lazily by the `Manager`. Since
now the profile is loaded by default when calling `verdi`, the backend
was always being loaded, even if not necessary, causing `verdi` and
therewith the tab-completion to grind to a halt.