Refactor link validation to improve efficiency and scaling
The slowness of the link validation stems from the fact that to validate
the indegree and outdegree of a proposed link, all incoming and outgoing
links are retrieved including the neighboring node instance. However, to
determine whether the new links violates the uniqueness constraint only
the UUID of the node is necessary and only a sub set of existing nodes
needs to checked, depending on the link type.

The number of nodes that are checked is reduced by using more specific
queries for each case instead of getting all neighboring nodes and doing
the check on the python level. This fixes the super-linear scaling of
the previous solution.

The `get_stored_link_triples` now takes the argument `only_uuid` which
can be used to project only the node UUID which prevents having to load
the entire instance.