Move auto import group outside django transaction (#3246)
The creation of the automatic import group for the Django backend did
not include newly created/imported Nodes.
By moving this (final) action outside the `with transaction.atomic()`
block, all Nodes are correctly included.

This was not caught by the tests, because while the node UUIDs that
should be present in the Group were checked, the Group.nodes generator
was used. If there are no Nodes in the group, it simply yields nothing,
i.e. nothing essentially fails. Now the Group.count() function is used
to make sure there is the correct number of Nodes in the Group.