Allow to construct `SinglefileData` nodes from binary files (#3285)
The `SinglefileData.set_file` method was opening the target file in
non-binary mode with default utf-8 encoding with no way for the user to
override it through the constructor. This would fail when a binary file
was passed. Since the purpose of this operation is just to copy the file
wholesale to the repository this should simply be done in binary mode to
copy the bytes over.