Fix the identifier format of the caching configuration file
The old caching configuration file implementation expected the module
path of node sub classes that should be either enabled or disabled for
caching, for example:

    `aiida.orm.node.calculation.job.TemplatereplacerCalculation`

The old caching implementation in the `Node.store` method would
then compare these directives to the node type of the current node being
stored to determine whether it should be considered for caching. This
method broke after the separation of the "process" and "node" concepts.
Since calculation job classes, for example, are now no longer a sub
class of `Node`, but of `Process`, the corresponding node that will be
created will all be `CalcJobNode` instances, regardless of the exact
process sub class. That particular bit of information is stored as its
entry point string in the `process_type` column of the node. For example:

    `aiida.calculations:templatereplacer`

This entry point string is now also the identifier that determines the
caching behavior and should be used in the configuration file.