Fixing the drop in test coverage reported (#3326)
Since removing the CWD from the `sys.path` in a recent commit, the test
coverage reported by coveralls dropped. This is due to the `pytest-cov`
module that was being used which explicitly does not include it, unlike
when running `coverage` directly. This caused the root directory of the
machine to be included and the coverage to drop. Reverting to running
`coverage` directly and dropping `pytest-cov` restores the behavior.