Allow empty `enabled` and `disabled` keys in caching configuration (#3330)
If the `enabled` or `disabled` key is specified in the caching
configuration but contains no values, or all values are commented, the
parsed dictionary will have `None` as the value for the corresponding
key. This was causing an exception in`aiida.manage.caching._get_config`
because it was always iterating over it without checking that it exactly
contained an iterable. Also `get_use_cache` was affected.