Fix unspecific flush in `DjangoNode` causing data override (#3344)
The methods of `DjangoNode` related to the mutation of attributes and
extras were indirectly causing a flush of the entire node state instead
of flushing only the property they were mutating. For example, changing
the extras, would also cause the attributes to be flushed. This could
lead to problems, if the attributes had been changed in another instance
and flushed to the database, because the original instance will not
refresh the state of the attributes when flushing its state, these
changes will be overridden.

The `SqlaNode` properly avoided this problem by "flagging" the changed
property before calling the flush. This is done by a SqlAlchemy utility
function which is not present for Django. Instead we pipe this through
to the `DbNode.save` method which supports the `updated_fields` argument
to target only specific model fields to be flushed.