Ensure exceptions when launching daemon runners are logged (#3348)
Since the daemon workers launched by the Circus daemonizer can except
before the logging is configured `aiida.engine.daemon.runner.start_daemon`
the stderr stream of those sub processes needs to be also forwarded by
the watchers for them to end up in the daemon log file. If not these
stack traces are lost forever and the user won't notice that there is
something wrong. Especially because `verdi daemon status` will actually
show the daemon to be running with an active worker, it is just that it
is a single worker everything, dying and immediately being revived.